Moved S3 services under S3 namespace
